split comm events to a separate module
