+++ extensions/positron-r/amalthea/crates/amalthea/src/comm/mod.rs
@@ -7,4 +7,5 @@
 
 pub mod comm_channel;
 pub mod comm_manager;
+pub mod event;
 pub mod lsp_comm;
